Split off the reference drivers repo implementation
